+++ libs/chromosome/chromosome_decoder.py
